Refactor: 친구 선택 위젯 분리 #151
