+++ lib/features/place/presentation/models/place_friend_profile.dart
@@ -1,0 +1,11 @@
+class PlaceFriendProfile {
+  const PlaceFriendProfile({
+    required this.id,
+    required this.name,
+    this.imageAsset,
+  });
+
+  final String id;
+  final String name;
+  final String? imageAsset;
+}
